feat: handle files in one level

+++ src/services/LibraryService.ts
@@ -14,6 +14,7 @@ import database from '../database';
 import { TYPES } from '../ContainerTypes';
 import { IStorageService } from '../interfaces/IStorageService';
 import { ILoggerService } from '../interfaces/ILoggerService';
+import moment from 'moment-timezone';
 
 @injectable()
 export class LibraryService {
@@ -193,7 +194,7 @@ export class LibraryService {
                 where user_id=? and active=true and key like ?
             ) as filtroKey) as filtro
       where ss.id_library_item = filtro.id_library_item
-      returning ss.id_library_item, ss.key, ss.type, filtro.old_key;
+      returning ss.id_library_item, ss.key, ss.type, filtro.old_key, ss.original_filename, ss.source_path;
       `,
           [destinationPath, origin, user_id, `${origin}%`],
         )
@@ -278,7 +279,7 @@ export class LibraryService {
                     where user_id=? and active=true and key like ?
                 ) as filtroKey) as filtro
       where ss.id_library_item = filtro.id_library_item
-      returning ss.id_library_item, ss.key, ss.type;
+      returning ss.id_library_item, ss.key, ss.type, ss.original_filename, ss.source_path;
       `,
           [folderPath, user_id, `${folderPath}/%`],
         )
@@ -319,6 +320,7 @@ export class LibraryService {
           type: item.type,
           is_finish: item.is_finish,
           thumbnail: item.thumbnail || null,
+          source_path: item.source_path,
         })
         .returning('*');
       return objects[0];
@@ -402,8 +404,9 @@ export class LibraryService {
               break;
             default: // deprecated old part
               if (options.withPresign) {
+                const originalFile = itemDb.source_path || itemDb.key;
                 const { url } = await this._storage.GetPresignedUrl({
-                  key: `${user.email}/${itemDb.key}`,
+                  key: `${user.email}/${originalFile}`,
                   type: StorageAction.GET,
                 });
                 fileUrl = url;
@@ -474,6 +477,7 @@ export class LibraryService {
           thumbnail: itemTemp.thumbnail,
           url: '',
           synced: itemTemp.synced,
+          source_path: itemTemp.source_path,
         };
         break;
       case LibraryItemOutput.DB:
@@ -499,13 +503,18 @@ export class LibraryService {
           is_finish: itemApi.isFinished,
           thumbnail: itemApi.thumbnail,
           synced: itemApi.synced !== undefined ? itemApi.synced : undefined,
+          source_path: itemApi.source_path,
         };
         break;
     }
     return parsed;
   }
 
-  async GetObject(user: User, path: string): Promise<LibraryItem> {
+  async GetObject(
+    user: User,
+    path: string,
+    appVersion?: string,
+  ): Promise<LibraryItem> {
     try {
       const cleanPath = path.replace(`${user.email}/`, '');
       const objectDB = await this.dbGetLibrary(user.id_user, cleanPath);
@@ -513,16 +522,32 @@ export class LibraryService {
       if (!itemDb) {
         throw Error('Item not found');
       }
-      const { url, expires_in } = await this._storage.GetPresignedUrl({
-        key: `${user.email}/${itemDb.key}`,
-        type: StorageAction.GET,
-      });
       const libObj = (await this.ParseLibraryItemDbB(
         itemDb,
         LibraryItemOutput.API,
       )) as LibraryItem;
-      libObj.url = url;
-      libObj.expires_in = expires_in;
+      let fileUrl = null;
+      switch (appVersion) {
+        case '2023-10-29':
+        case 'latest':
+          fileUrl =
+            parseInt(itemDb.type) === parseInt(LibraryItemType.BOOK)
+              ? `${process.env.PROXY_FILE_URL}/${encodeURIComponent(
+                  itemDb.key,
+                )}`
+              : null;
+          break;
+        default: // deprecated old part
+          const originalFile = itemDb.source_path || itemDb.key;
+          const { url, expires_in } = await this._storage.GetPresignedUrl({
+            key: `${user.email}/${originalFile}`,
+            type: StorageAction.GET,
+          });
+          fileUrl = url;
+          libObj.expires_in = expires_in;
+          break;
+      }
+      libObj.url = fileUrl;
       return libObj;
     } catch (err) {
       this._logger.log({
@@ -548,9 +573,12 @@ export class LibraryService {
         exactly: true,
       });
       let itemDb = objectDB[0];
+      libObj.source_path = `${process.env.ROOT_FOLDER}/${moment().format(
+        'YYYYMMDDHHmmss',
+      )}_${libObj.original_filename}`;
       if (itemDb) {
         const fileExists = await this._storage.fileExists({
-          key: `${user.email}/${relativePath}`,
+          key: `${user.email}/${itemDb.source_path || itemDb.key}`,
         });
         if (fileExists === true) {
           return null;
@@ -558,12 +586,7 @@ export class LibraryService {
       } else {
         itemDb = await this.dbInsertLibraryItem(user.id_user, libObj);
       }
-
-      // S3 needs the forward slash to create an empty folder
-      const resourcePath =
-        parseInt(libObj.type) === parseInt(LibraryItemType.BOOK)
-          ? `${user.email}/${relativePath}`
-          : `${user.email}/${relativePath}/`;
+      const resourcePath = `${user.email}/${libObj.source_path}`;
 
       const { url, expires_in } = await this._storage.GetPresignedUrl({
         key: resourcePath,
@@ -601,7 +624,7 @@ export class LibraryService {
 
       for (let index = 0; index < deletedObjects.length; index++) {
         const item = deletedObjects[index];
-        const sourceKey = `${user.email}/${item.key}`;
+        const sourceKey = `${user.email}/${item.source_path || item.key}`;
         await this._storage.deleteFile({
           sourceKey: `${sourceKey}${
             parseInt(item.type) === parseInt(LibraryItemType.BOOK) ? '' : '/'
@@ -765,6 +788,7 @@ export class LibraryService {
             },
             trx,
           );
+
           destinationDB = await this.dbGetLibrary(
             user.id_user,
             destinationPathFolder,
@@ -810,16 +834,36 @@ export class LibraryService {
 
       for (let index = 0; index < dbMoved.length; index++) {
         const fileMoved = dbMoved[index];
-        const suffix =
+        if (
+          !fileMoved.source_path &&
           parseInt(fileMoved.type) === parseInt(LibraryItemType.BOOK)
-            ? ''
-            : '/';
-        const sourceKey = `${user.email}/${fileMoved.old_key}${suffix}`;
-        const targetKey = `${user.email}/${fileMoved.key}${suffix}`;
-        await this._storage.moveFile({
-          sourceKey,
-          targetKey,
-        });
+        ) {
+          const suffix =
+            parseInt(fileMoved.type) === parseInt(LibraryItemType.BOOK)
+              ? ''
+              : '/';
+          const sourceKey = `${user.email}/${fileMoved.old_key}${suffix}`;
+          const original_filename = `${
+            process.env.ROOT_FOLDER
+          }/${moment().format('YYYYMMDDHHmmss')}_${
+            fileMoved.original_filename
+          }`;
+          const targetKey = `${user.email}/${original_filename}`;
+          const isMoved = await this._storage.moveFile({
+            sourceKey,
+            targetKey,
+          });
+          if (isMoved) {
+            await trx('library_items')
+              .update({
+                source_path: original_filename,
+              })
+              .where({
+                user_id: user.id_user,
+                key: fileMoved.key,
+              });
+          }
+        }
       }
 
       await trx.commit();
@@ -868,23 +912,49 @@ export class LibraryService {
           (preFile) => preFile.id_library_item === fileMoved.id_library_item,
         );
 
-        if (prevFile) {
+        if (
+          prevFile &&
+          !fileMoved.source_path &&
+          parseInt(fileMoved.type) === parseInt(LibraryItemType.BOOK)
+        ) {
           const suffix =
             parseInt(prevFile.type) === parseInt(LibraryItemType.BOOK)
               ? ''
               : '/';
           const sourceKey = `${user.email}/${prevFile.key}${suffix}`;
-          const targetKey = `${user.email}/${fileMoved.key}${suffix}`;
-          await this._storage.moveFile({
+          const original_filename = `${
+            process.env.ROOT_FOLDER
+          }/${moment().format('YYYYMMDDHHmmss')}_${
+            fileMoved.original_filename
+          }`;
+          const targetKey = `${user.email}/${original_filename}`;
+          const isMoved = await this._storage.moveFile({
             sourceKey,
             targetKey,
           });
+          if (isMoved) {
+            await this.db('library_items')
+              .update({
+                source_path: original_filename,
+              })
+              .where({
+                user_id: user.id_user,
+                key: fileMoved.key,
+              });
+          }
         }
       }
-
-      await this._storage.deleteFile({
-        sourceKey: `${user.email}/${folderDB[0].key}/`,
+      const folderKey = `${user.email}/${
+        folderDB[0].source_path || folderDB[0].key
+      }/`;
+      const folderExist = await this._storage.fileExists({
+        key: folderKey,
       });
+      if (folderExist) {
+        await this._storage.deleteFile({
+          sourceKey: folderKey,
+        });
+      }
 
       await trx.commit();
       return true;
@@ -941,8 +1011,9 @@ export class LibraryService {
           break;
         default: // deprecated old part
           if (options.withPresign) {
+            const originalFile = item.source_path || itemDb.key;
             const { url, expires_in } = await this._storage.GetPresignedUrl({
-              key: `${user.email}/${itemDb.key}`,
+              key: `${user.email}/${originalFile}`,
               type: StorageAction.GET,
             });
             item.url = url;
@@ -1099,6 +1170,7 @@ export class LibraryService {
             key: itemDb[0].key,
             old_key: itemDb[0].key,
             type: itemDb[0].type,
+            original_filename: itemDb[0].original_filename,
           },
         ];
       }
@@ -1126,16 +1198,36 @@ export class LibraryService {
       );
       for (let index = 0; index < dbMoved.length; index++) {
         const fileMoved = dbMoved[index];
-        const suffix =
+        if (
+          fileMoved.original_filename.indexOf(process.env.ROOT_FOLDER) !== -1 &&
           parseInt(fileMoved.type) === parseInt(LibraryItemType.BOOK)
-            ? ''
-            : '/';
-        const sourceKey = `${user.email}/${fileMoved.old_key}${suffix}`;
-        const targetKey = `${user.email}/${fileMoved.key}${suffix}`;
-        await this._storage.moveFile({
-          sourceKey,
-          targetKey,
-        });
+        ) {
+          const suffix =
+            parseInt(fileMoved.type) === parseInt(LibraryItemType.BOOK)
+              ? ''
+              : '/';
+          const sourceKey = `${user.email}/${fileMoved.old_key}${suffix}`;
+          const original_filename = `${
+            process.env.ROOT_FOLDER
+          }/${moment().format('YYYYMMDDHHmmss')}_${
+            fileMoved.original_filename
+          }`;
+          const targetKey = `${user.email}/${original_filename}`;
+          const isMoved = await this._storage.moveFile({
+            sourceKey,
+            targetKey,
+          });
+          if (isMoved) {
+            await this.db('library_items')
+              .update({
+                original_filename,
+              })
+              .where({
+                user_id: user.id_user,
+                key: fileMoved.key,
+              });
+          }
+        }
       }
 
       await trx.commit();
